Change NewAuthor to allow updates
